+++ src/assets/css/header.css
@@ -1,0 +1,3 @@
+.tech{
+    width: 100%;
+}
