+++ src/components/main.js
@@ -1,0 +1,10 @@
+import React, {Component} from 'react';
+import Header from './header.js'
+
+class Main extends Component{
+    render(){
+        return(
+            <Header/>
+        )
+    }
+}
